requirements.txt
attrs==25.3.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
networkx==3.5
referencing==0.36.2
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
typing_extensions==4.14.0
